requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
astroid==2.3.1
attrs==19.2.0
black==19.3b0
Click==7.0
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-graphql-jwt==0.2.2
django-graphql-playground==1.0
djangorestframework==3.10.3
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
promise==2.2.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.2
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2
